Repeatables put multiple Add buttons into dropdown
For repeatable items that have multiple item types that can be added, the user interface can become too crowded with "Add Item" links.

This commit detects if there are more than three "Add Item" links, and if so it creates a dropdown where the user can select the item type, and a single "Add" button. Clicking the add button will add the item type that is currently selected in the dropdown.