RTE add readOnly style parameter
Added ability to set a style definition to readOnly, which means once set the content inside the style cannot be modified by the user. The user also cannot place the cursor inside the style, but the dropdown appears when the cursor is next to the style so the user can click "clear" to remove the style.