RTE table check for context
Allow a style definition to be set for the "table" element. The style definition is not actually used to create the table style (since table is a special case like an enhancement); however, the style definition can contain a context array like other styles. As the user moves the cursor, this context will be checked to determine when the table toolbar button should be disabled.

This commit also includes some logic so if the backend specifies a RichTextElement for "table" it will not be added to the toolbar (since tables are not handled like other styles).