Fixes sticky elements jerking up/down on children resize.
Because sticky-kit uses CSS bottom property to implement its bottoming
logic, there was a sudden jerk whenever an element got bigger. This
happened because when an element is bottom aligned, it grew UP on
resize, and sticky-kit corrected that on recalc by shifting everything
back. This commit changes sticky-kit to use CSS top property for the
same functionality to reduce this jitter.