RTE update handsontable and add table attributes
Update the handsontable bower version, to fix a bug where meta data was not moving with the cells to which it is attached, which is needed for the table attributes.

For the table HTML parsing and HTML export, add support for table / tr / td attributes. This is needed so we can later add a RichText element that sets attributes.