RTE add UI for table/tr/td attribute popup forms
Gives ability to specify a RichText element on the backend using tag table, tr, or td and display an "inline enhancement" style popup form for editing attributes on table, table row, and table cell.

If such elements are specified on the backend, then the RTE will add the following items to the dropdown menu that appears when you click on a table cell: Edit Table Attribues, Edit Row Attributes, Edit Cell Attributes.

The popup form should appear letting you edit the attributes for the table, tr, or td element. However, the backend cannot set the innerHTML of these elements (as it can with other element types) - only attribute setting is allowed.