RTE bugfix for table toolbar context
Modify the toolbar code so any RichText element that is tag=table will create a new toolbar button. Previously the table element was ignored and it was assumed that the table button would be added statically; however, due to the way we handle context root for each field, we need ability for the backend to specify which fields will have a toolbar button, and which fields should omit the toolbar button.