RTE add dropdown to comments to show user and time
We are saving the user name and timestamp as attributes on the comments. Currently there is no way for the user to see this information. This commit adds a dropdown that appears when the cursor is over the comment, the dropdown shows the username, the date, and the time at which the comment was originally created.