RTE add style to force scrollbar display for table
Within an RTE table, if a cell can expand and push other cells to the right. The user must scroll the div that contains the table in order to access the other columns.

However, in Mac OS X, scrollbars are hidden by default until the user tries to scroll. This can lead to some confusion since the user does not necessarily know that the area is scrollable.

This commit adds some styles so the scrollbar will always appear (when the table is scrollable) and the user will know the area is scrollable.