RTE bugfix for enhancement move
Fix a problem when moving enhancement down when the enhancement is already on the last line. The intended behavior is to add a new line to the end of the content, then move the enhancement down.

Currently an error occurs when the new line is added due to a missing argument in the call to CodeMirror.

Also, another recent fix (to adjust the position of an enhancement when the user presses enter at the beginning of the line) was causing problems when moving an enhancement down. This commit adds some checking to determine if the user pressed enter vs pressing the down arrow button for the enhancement.