RTE mark changed only after user input
Since the RTE must parse the input HTML, in some cases the output will not match the input, even in cases where the user has not interacted with the editor to change the content.

This commit sets a "changed" flag on the editor only after the user has performed some modification to the editor content. If the content has not been changed by the user, then the textarea will not be updated with the editor output.

This will prevent the input from being marked as "changed" until the user modifies the content, so the user will not be prevented from leaving the page.

Also it means that if the user publishes the page, the original value in the textarea will be retained (and not the modified output from the editor).