RTE improve change detection timing
This commit changes the rteChange event so it is not debounced at the source of CodeMirror "changes" events. Instead the rteChange event will fire immediately whenever there is a CodeMirror "changes" event. It is the responsibility of any code listening to the rteChange event to properly throttle/debounce so that the code does not run too often and cause performance issues (luckily it appears that all other code listening for rteChange event already does this, so the initial debounce was not needed).

By removing this debounce at the source, we are able to prevent some timing issues that were complicating marking the RTE content as "changed". Now we can set the initial HTML content of the RTE, and accurately listen for rteChange event to tell when the user has made a change to the editor.

Previously we were getting delayed change events from CodeMirror after we set the initial content of the editor, and this made it impossible to determine if the change was coming from the initial content or from a user's change. This fix lets us remove a setTimeout() that was used previously and was not an ideal solution since it could potentially break down on a slow system.

Now the change events should be guaranteed to occur synchronously, so after setting the initial HTMl we can listen for change events to determine when the user actually updated the content.