Repeatable bugfix click edits wrong slide
Due to a recent change we added (when switching to gallery view automatically edit the first slide if no slide has been selected, commit d76701a0b481f628a73ecb9db9db4f2dbb6735d8), clicking "Edit" on a slide was not selecting the correct slide to edit, instead it was always editing the first slide.

This commit changes the order in which the active slide is set to fix this issue.