RTE allow style to toggle off when it has context
If a style definition has context rules, currently the toolbar cannot be used to toggle off that style. Since typically the context rules might not let you use a style within itself, the toolbar button gets disabled, and therefore the user cannot click the button to toggle off the style.

This commit adds a check when every character in the range has the style, and the style is toggleable, then do not disable the button (since clicking the button would only remove the style).