RTE adjust enhancement to work with other lineWidgets
There is some specialized code for enhancements that handles pressing enter at the start of the line. This code looks for any lineWidget on the line; however, since we just added another use of lineWidgets, this code was finding things that were not enhancements. Added a flag to the enhancement lineWidget so we can detect when we are dealing with an enhancement vs some other type of lineWidget.