RTE float enhancements fix toolbar and dropdown
In the RTE if you create an enhancement (old style) and make it float to the right or left, then the toolbar buttons did not work if they wrapped to two lines. In addition, the image size dropdown did not work because the items in the dropdown were not getting mouse events.

This commit changes the enhancement toolbar so the buttons will not wrap, so the buttons will always get mouse events. This might extend the size of the floated enhancement slightly, but not much wider than it already was.

For the image size dropdown menu, it was not possible to use the easy CSS-based dropdown, so a javascript solution was implemented to overlay the dropdown when the mouse hovers the "image size" toolbar button.