RTE use context rules for html element order
Currently the RTE orders HTML elements only using a last-in-first-out order - if you select some text and mark it as bold, and then italic, you end up with <I><B> as the HTML.

However in some cases due to context rules this results in invalid HTML output. For example, if element B is only allowed within element A, and the user marks the same text with both elements, we need to ensure that we output <A><B> and not the other way around.

This commit slightly changes the sorting of the code mirror marks - it only changes the order when multiple styles start at the same character, and only if the styles have context rules that require a specific order.