RTE double click for edit popup
For a rich text element / inline enhancement, double click in the editor should popup the edit form. If multiple marks exist at the point of the double click, then the first mark found should be edited. Alternative it to single click and get a dropdown menu showing all the available marks and clicking "Edit" on the appropriate one.
Note CodeMirror provides a dblclick event but it does not work reliably, so we have simulated double clicking using timeouts and the mousedown event.