RTE bugfix dynamic-placeholder not showing
When using data-dyanmic-placeholder, the palceholder was not showing on initial load of the RTE. This is because the placeholder is set after the RTE initializes. This commit adds a placeholderUpdate event to the state.js code that sets the dynamic placeholder, so the RTE can listen for this event to know when the placeholder is available.