RTE add support for RichText elementContainer
Adding support for elements like list elements to be specified in a RichText specification on the backend. Added to the RTE frontend code, support for elementContainer property read from the RICH_TEXT_ELEMENTS global variable. Note this will require additional work on the backend to add the elementContainer property to RichText elements and to output that value as part of the RICH_TEXT_ELEMENTS global variable.