RTE history/undo/redo fixes
This commit fixes a problem with "redo" where the redo action was adding more actions to the history (which has the effect of clearing the redo history at that point forward, so you couldn't continue to redo other actions).

Also fixes a problem where performing a redo resulted in different styles - if you did a redo next to another mark, the other mark was extended to to cover all the redo content. To fix this, before content is added, the marks at the point are "split" when necessary to ensure the marks are not surrounding the new content.