RTE history/undo/redo fix for mark positions
Modify the information that is saved for each change, so the mark position is saved at the point the change is made. This fixes a problem where undo was modifying the marks and making them smaller than they should have been, then when a redo was performed the text inserted back in was not being styled as expected.