RTE undo/redo prevent empty mark in hsitory
In some cases like inline enhancements, we set a zero-length mark as a placeholder before text is inserted in the mark. This is causing problems for the undo history since it can create a separate history event. This commit ensures that creating a zero-length mark will not create a history event.