RTE history/undo/redo fixes for redo
Removed code from the redo operation that was duplicating marks. Should not be necessary for redo to recreate marks because the undo operation should be setting up marks as they were before text was inserted.
Note there are still problems with inline enhancements that insert HTML.
Also added "atomic" parameter to be saved with marks.