RTE prevent preview fetch unless style has flag
To improve performance, avoid fetching preview data on all line styles, instead fetch preview data only for styles that support it.
This commit adds a check to the updatePreview() function, so only styles that have a previewable:true flag are fetched. This is the FE portion of this code only, it is assumed that the style definitions in the global RICH_TEXT_ELEMENTS variable will be modified to set a previewable flag for the styles that support preview.