RTE undo/redo update marks in history
In some cases a link to the the same mark can be saved in multiple history events, and if you undo the mark can get recreated. This commit ensures that when a mark is recreated, the other history events will automatically update to get the newest version of the mark.