RTE bugfix table not saving changes
There is a bug where changes to an existing table are not saved when the page is published. This occurred because the RTE did not recognize that changes were made, and did not update the HTML output.
This commits adds an event to the table editor to indicate when the table is changed, and an event listener to the RTE.