RTE indent continued development
Modified indent level to allow a zero-indent level (to allow non-list lines to be indented at same level inside a list, or to break out of a list).
Modified the initListListeners function to properly handle indent levels when user presses enter at end of a line to create a new list item.