RTE nested list improve tab and backspace
Improve the behavior of Tab, Shift-Tab, and Backspace when dealing with list items. Only indent the list when the cursor is at the start of a line. Only indent additional lines if the previous line is a list item. Pressing Backspace at the start of the line should reduce the indentation level if the list is indented, or remove the list style if at the minimum indentation.