RTE undo changes for inline enhancement popup
This commit makes changes to the inline enhancement popup (enhancement.jsp) for the case when the inline enhancement updates the body of the element, so the changes will be included in the undo history.

There are still some problems when editing the same enhancement multiple times, but this is an improvement over the previous code which did not support undo history  at all.