RTE make block styles start on new line
In cases where external HTML is imported, if there is text that does not end the line, then a block-style element is encountered, the previous text was being included inside the block element. This commit adds a check to add a newline when necessary to ensure the text is separate from the block element. Note that this will also cause the exported HTML to add a <BR> element before the block element, but that is standard for the HTML that the RTE creates.