RTE make tables and enhancements start on new line
In cases where external HTML is imported, if there is text that does not end the line, then a table or enhancement is encountered, the previous text was being moved after the table or enhancement. This commit adds a check to add a newline when necessary to ensure the text is not moved. Note that this will also cause the exported HTML to add a <BR> element before the table or enhancement, but that is standard for the HTML that the RTE creates.