RTE paste table content is deleting whitspace
In some where a table contains formatting with the table cells, whitespace was getting deleted when it should not. This commit adjusts the code that was deleting leading/trailing whitespace so it occurs only at the start or end of the table cell (instead of text nodes that might be in the middle of other table cell content).