RTE performance improvement for list items
When a list item style is set on a line, the RTE also checks to see if there are other styles that need to be removed. For example, if you set a line to a bulleted list, it must remove the numbered list style to avoid conflicts.

When doing so, the RTE was also calling refresh() on the editor for each style removed, and for documents with many list items, this caused the editor to refresh the DOM too often, and performance suffered.

This commit wraps the refresh() function within a debounced timer, so it will not be called too frequently when importing HTML.