Search carousel previous not working
When the search results contain greater than 10 results, when you get to the last “page” of results in the carousel, sometimes the “previous” button does not work to fetch previous results for the carousel tiles. This commit corrects the logic in the carousel so it will trigger both a carouselEnd and carouselStart event when necessary, so the search results knows to fetch more results.