RTE paste from MSWord paragraph spacing
Currently pasting from MSWord, paragraphs are converted into single lines within the RTE with no space between paragraphs. This works fine for the default word style, but in some cases where paragraph styles have been applied, the Word file appears with spaces between the paragraphs. In those cases when pasting into the RTE, we lose the space between paragraphs and it is confusing to the user. This commit checks the pasted HTML that comes from word, and if the paragraph has a margin-bottom style attribute, then we assume a blank line should be placed after the paragraph, so the content in the RTE will more closely reflect the Word document.