RTE track changes undo fix inline cleanup
The calls to inlineCleanup were modifying the undo history, when used with track changes. This commit removes the calls to inline cleanup during normal editing, and moves that logic to be performed on-the-fly during the toHTML export. So instead of actually changing CodeMirror marks when the marks were adjacent to each other, the marks are left as-is, and the styles are combined during the HTML output only, and thus the the undo history is not modified.