RTE after link edit focus on editor with no append
Feature request for the RTE after editing a link the focus should return to the editor. Also the cursor should be placed after the link, and immediately typing more text should not expand the link. Note that if you move the cursor, then return it after the link, then typing will expand the link (this is standard behavior for marked regions, such as bold, etc.)