RTE move context checking out of toolbar handler
Currently the RTE context checking is performed by the toolbar update function, which means it can’t be used in other areas such as the keyboard shortcuts. This commit starts the work by separating out context checking into its own function and modifying the toolbar update function to use the more generic function. Then in subsequent updates we can add context checking to the keyboard shortcuts, so users will not be able to apply styles where the context does not allow.