RTE track changes paste over undo history
When track changes is turned on, if you copy some text, then select another range of text and paste over it, then perform an undo, the editor is not returning to the previous state. This commit changes some of the history logic so it correctly detects the changes that a track history paste makes.