Add field parameters to inline editor links
Parse the source to find BrightspotCmsFieldAccess comments which indicate which fields are being used for each object. Then in the inline edit link for each object, include url parameters so only the fields that are used will be shown in the edit form. So instead of the url “edit.jsp?id=123” it would include field info like “edit.jsp?id=123&f=firstname&f=lastname”.