Inline editor add listener for update message
The backend can use postMessage to send a message to the parent window, to let it know that that the content has been updated and the page should be reloaded to show the new content. Also adds iframe resizer code, so the backend can include /script/iframeResizer.contentWindow.js to set up auto resizing in the iframe.