RTE bugfix for missing end tag on comments etc
If an inline style was at the end of the line, and the user pressed enter to create a new line, the inline style carries over to the new line. However, on HTML output, the style crossing the line boundary was causing a bug where the ending HTML tag was omitted. Result was the style would be applied to the rest of the content. This commit updates the logic that checks CodeMirror marks, to account for marks that cross line boundaries.