BSP-2278 inlineEdit iframe padding and resize
This commit has several changes:
Modify the inlineEdit page so it drops some javascript on the page that will add a class name to the body element. This lets us override some global CSS, but only for this page.
Adds some CSS to override the padding on certain elements, to eliminate extra padding on the inlineEdit page which is in an iframe.
Modify the sticky javascript code so it does not make the toolbar sticky when the page is in an iframe. This was causing problems on the inlineEdit page.
Modify the iframe resize javascript so it will work as expected when certain dropdowns appear on the page and extend beyond the body element.