BSP-2622 add FE controls for inline text field
Work in progress to add inline contenteditable editing for text fields.
Text fields are found on the live page and when mouse hovers over them the cursor changes (to the “cell” pointer). Clicking on the text makes the text editable and a save and cancel button appears.
Note this is a work in progress as the backend service still needs to be created, however an existing backend page is temporarily used and you can actually change the text value.