Move preview javascript out of jsp
The javascript that controls the preview functionality is embedded into the JSP. This commit moves it into a separate javascript file that will be included only on the edit page. Also since the javsacript had some embeded JSP code, put those values into a separate javscript object that is then referenced in the javascript.