Preview field mapping fix for initial scroll pos
Fix for the initial scroll postion of the preview window, when activating the field mapping.

Problem: Field mapping visualization does not outline the correct area in the preview window if the preview window is not scrolled to the top

Steps to reproduce:
Display the preview and pin it.
Scroll down in the preview content
In the CMS edit form click the "eye" icon next to a field. The field preview visualization appears.

Expected result:
The field mapping visualization should outline the field in the preview window.

Observed result:
The field mapping visualization appears as it would if the preview window was scrolled to the top, and the outline box does not match the actual scrolled position of the field.